refactor: consolidate path utilities into new agtrace-core infrastructure crate (fixes #19)
