+++ crates/agtrace-core/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "agtrace-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+agtrace-types.workspace = true
+sha2.workspace = true
+dirs = "5.0"
+
+[dev-dependencies]
+tempfile.workspace = true
