+++ crates/agtrace-core/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod path;
+
+pub use path::*;
