+++ crates/agtrace-providers/src/gemini/discovery.rs
@@ -1,6 +1,6 @@
 use crate::traits::{LogDiscovery, ProbeResult, SessionIndex};
 use crate::{Error, Result};
-use agtrace_types::project_hash_from_root;
+use agtrace_core::project_hash_from_root;
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use walkdir::WalkDir;
