+++ crates/agtrace-runtime/src/config.rs
@@ -3,53 +3,16 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::PathBuf;
 
-// TODO: Path utilities are currently scattered across crates (technical debt).
-// This function is temporarily in runtime/config.rs.
-// Related utilities like `discover_project_root()` are in agtrace-types/src/util.rs
-// but types should only contain schemas, not utilities.
-// See: https://github.com/lanegrid/agtrace/issues/19
-
 /// Resolve the workspace data directory path based on priority:
 /// 1. Explicit path (with tilde expansion)
 /// 2. AGTRACE_PATH environment variable (with tilde expansion)
 /// 3. XDG data directory (recommended default)
 /// 4. ~/.agtrace (fallback for systems without XDG)
 pub fn resolve_workspace_path(explicit_path: Option<&str>) -> Result<PathBuf> {
-    // Priority 1: Explicit path
-    if let Some(path) = explicit_path {
-        return Ok(expand_tilde(path));
-    }
-
-    // Priority 2: AGTRACE_PATH environment variable
-    if let Ok(env_path) = std::env::var("AGTRACE_PATH") {
-        return Ok(expand_tilde(&env_path));
-    }
-
-    // Priority 3: XDG data directory (recommended default)
-    if let Some(data_dir) = dirs::data_dir() {
-        return Ok(data_dir.join("agtrace"));
-    }
-
-    // Priority 4: Fallback to ~/.agtrace (last resort for systems without XDG)
-    if let Some(home) = std::env::var_os("HOME") {
-        return Ok(PathBuf::from(home).join(".agtrace"));
-    }
-
-    // This should never happen, but provide a working directory fallback
-    Err(Error::Config(
-        "Could not determine workspace path: no HOME directory or XDG data directory found"
-            .to_string(),
-    ))
-}
-
-/// Expand tilde (~) in paths to the user's home directory
-fn expand_tilde(path: &str) -> PathBuf {
-    if let Some(stripped) = path.strip_prefix("~/")
-        && let Some(home) = std::env::var_os("HOME")
-    {
-        return PathBuf::from(home).join(stripped);
-    }
-    PathBuf::from(path)
+    agtrace_core::resolve_workspace_path(explicit_path).map_err(|e| match e {
+        agtrace_core::path::Error::Io(io_err) => Error::Io(io_err),
+        agtrace_core::path::Error::Config(msg) => Error::Config(msg),
+    })
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
