+++ crates/agtrace-runtime/src/runtime/supervisor.rs
@@ -1,8 +1,8 @@
 use crate::runtime::events::{DiscoveryEvent, WorkspaceEvent};
 use crate::{Error, Result};
+use agtrace_core::project_hash_from_root;
 use agtrace_index::Database;
 use agtrace_providers::ProviderAdapter;
-use agtrace_types::project_hash_from_root;
 use notify::{Event, EventKind, PollWatcher, RecursiveMode, Watcher};
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
