+++ crates/agtrace-types/Cargo.toml
@@ -12,7 +12,6 @@ serde.workspace = true
 serde_json.workspace = true
 chrono.workspace = true
 uuid.workspace = true
-sha2.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
